Crate cortex_a

Low level access to Cortex-A processors

This crate provides:

Safe wrappers around assembly instructions

For now, there's not much. I will update it gradually. If you want to contribute, feel free to reach out!

Minimum Supported Rust Version (MSRV)

This crate is guaranteed to compile on stable Rust 1.39 and up. It might compile with older versions but that may change in any new patch release.

Modules

asm
Miscellaneous assembly instructions

barrier

regs
Processor core registers