Module cortex_a::regs::ESR_EL1::EC

Exception Class. Indicates the reason for the exception that this register holds information about.

For each EC value, the table references a subsection that gives information about:

The cause of the exception, for example the configuration required to enable the trap.

The encoding of the associated ISS.

Incomplete listing - to be done.

Enums

Value
Exception Class. Indicates the reason for the exception that this register holds information about.

Constants

Bkpt32

BranchTarget

BreakpointCurrentEL

BreakpointLowerEL

Brk64

CLEAR

DataAbortCurrentEL

DataAbortLowerEL

HVC64

IllegalExecutionState

InstrAbortCurrentEL

InstrAbortLowerEL

PCAlignmentFault

PointerAuth

SET

SError

SMC64

SPAlignmentFault

SVC32

SVC64

SoftwareStepCurrentEL

SoftwareStepLowerEL

TrappedFP

TrappedFP32

TrappedFP64

TrappedLDCorSTC

TrappedMCRRorMRRC

TrappedMCRorMRC

TrappedMCRorMRC2

TrappedMRRC

TrappedMsrMrs

TrappedSve

TrappedWFIorWFE

Unknown

WatchpointCurrentEL

WatchpointLowerEL