Module cortex_a::regs::MIDR_EL1::Implementer

The Implementer code. This field must hold an implementer code that has been assigned by Arm. Assigned codes include the following:

Hex representation Implementer 0x00 Reserved for software use 0xC0 Ampere Computing 0x41 Arm Limited 0x42 Broadcom Corporation 0x43 Cavium Inc. 0x44 Digital Equipment Corporation 0x46 Fujitsu Ltd. 0x49 Infineon Technologies AG 0x4D Motorola or Freescale Semiconductor Inc. 0x4E NVIDIA Corporation 0x50 Applied Micro Circuits Corporation 0x51 Qualcomm Inc. 0x56 Marvell International Ltd. 0x69 Intel Corporation

Arm can assign codes that are not published in this manual. All values not assigned by Arm are reserved and must not be used.

Enums

Value
The Implementer code. This field must hold an implementer code that has been assigned by Arm. Assigned codes include the following:

Constants

Ampere

AppliedMicroCircuits

Arm

Broadcom

CLEAR

Cavium

DigitalEquipment

Fujitsu

Infineon

Intel

Marvell

MotorolaOrFreescale

NVIDIA

Qualcomm

Reserved

SET