Module cortex_a::regs::SPSR_EL1::V

Overflow condition flag.

Set to the value of the V condition flag on taking an exception to EL1, and copied to the V condition flag on executing an exception return operation in EL1.

Set to 1 if the last flag-setting instruction resulted in an overflow condition, for example a signed overflow on an addition.

Enums

Value
Overflow condition flag.

Constants

CLEAR

SET