Module cortex_a::regs::SPSR_EL2::M

AArch64 state (Exception level and selected SP) that an exception was taken from. The possible values are:

M[3:0] | State

0b0000 | EL0t 0b0100 | EL1t 0b0101 | EL1h 0b1000 | EL2t 0b1001 | EL2h

Other values are reserved, and returning to an Exception level that is using AArch64 with a reserved value in this field is treated as an illegal exception return.

The bits in this field are interpreted as follows:

M[3:2] holds the Exception Level.

M[1] is unused and is RES 0 for all non-reserved values.

M[0] is used to select the SP:

0 means the SP is always SP0.

1 means the exception SP is determined by the EL.

Enums

Value
AArch64 state (Exception level and selected SP) that an exception was taken from. The possible values are:

Constants

CLEAR

EL0t

EL1h

EL1t

EL2h

EL2t

SET