Module cortex_a::regs::ESR_EL1::EC

Exception Class. Indicates the reason for the exception that this register holds information about.

For each EC value, the table references a subsection that gives information about:

The cause of the exception, for example the configuration required to enable the trap.

The encoding of the associated ISS.

Incomplete listing - to be done.

Enums

Value
Exception Class. Indicates the reason for the exception that this register holds information about.

Constants

CLEAR

DataAbortCurrentEL

DataAbortLowerEL

HVC64

IllegalExecutionState

InstrAbortCurrentEL

InstrAbortLowerEL

PCAlignmentFault

SET

SMC64

SPAlignmentFault

SVC64

TrappedFP

TrappedFP64

TrappedMsrMrs

TrappedWFIorWFE

Unknown